Feat: Generate declaration files corretly - add --rootDir src option to avoid generating redundant declaration files - move Story.tsx files to fix error: ‘rootDir’ is expected to contain all source files.
